requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
pycodestyle==2.8.0
pydantic==1.8.2
PyJWT==1.7.1
python-dotenv==0.19.2
python-fity3==0.7
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
